add imgproc helpers
