+++ Server/core/vision/detectors/contour_detector.py
@@ -46,16 +46,14 @@ from ..config_defaults import (
 )
 from ..vision_utils import pct_on, despeckle
 from ..dynamic_adjuster import CannyConfig
+from .imgproc import (
+    _preprocess,
+    _color_gate,
+    _try_with_margins,
+    _draw_overlay,
+)
 
 NDArray = np.ndarray
-
-# ----------------------- small helpers -----------------------
-def _odd(k: int) -> int:
-    k = int(k)
-    return k if (k % 2 == 1) else k + 1
-
-def _clip01(x: float) -> float:
-    return float(max(0.0, min(1.0, x)))
 
 # ----------------------- configs -----------------------
 @dataclass
@@ -265,7 +263,7 @@ class ContourDetector:
             stamp = time.strftime("%Y%m%d_%H%M%S")
 
         # ----- Preprocess -----
-        proc, gray = self._preprocess(img)
+        proc, gray = _preprocess(img, self.proc)
 
         # ----- Dynamic adjuster (auto canny + rescue) -----
         edges, canny, t1, t2, life, used_rescue = self.adjuster.apply(gray)
@@ -283,7 +281,7 @@ class ContourDetector:
         color_cover = 0.0
         color_used = False
         if self.color.enabled:
-            color_mask = self._color_gate(proc)
+            color_mask = _color_gate(proc, self.color)
             color_cover = pct_on(color_mask)
             # sanity check: ignore if too small/too big
             if (color_cover < self.color.min_cover_pct) or (color_cover > self.color.max_cover_pct):
@@ -324,7 +322,7 @@ class ContourDetector:
             cv2.imwrite(os.path.join(save_dir, f"{stamp}_edges_patched.png"), edges2)
 
         # ----- Main selection -----
-        best, e_used = self._try_with_margins(edges2)
+        best, e_used = _try_with_margins(edges2, self.proc, self.morph_cfg, self.geo, self.w)
         if save_dir is not None:
             cv2.imwrite(os.path.join(save_dir, f"{stamp}_edges_used.png"), e_used)
 
@@ -341,7 +339,7 @@ class ContourDetector:
             )
 
         mask_final, info, chosen_ck, chosen_dk = best
-        overlay, center = self._draw_overlay(proc, info, mask_final)
+        overlay, center = _draw_overlay(proc, info, mask_final, self.color.enabled)
 
         if save_dir is not None:
             cv2.imwrite(os.path.join(save_dir, f"{stamp}_mask_final.png"), mask_final)
@@ -394,164 +392,6 @@ class ContourDetector:
                 return cv2.cvtColor(img_or_path, cv2.COLOR_GRAY2BGR)
             return img_or_path.copy()
         return None
-
-    def _preprocess(self, img: NDArray):
-        proc = cv2.resize(img, (self.proc.proc_w, self.proc.proc_h), interpolation=cv2.INTER_AREA)
-        gray = cv2.cvtColor(proc, cv2.COLOR_BGR2GRAY)
-        gray = cv2.GaussianBlur(gray, (_odd(self.proc.blur_k), _odd(self.proc.blur_k)), 0)
-        return proc, gray
-
-    def _run_morph(self, edges: NDArray, ck: int, dk: int, opening: bool = False) -> NDArray:
-        m = edges.copy()
-        if opening:
-            k0 = cv2.getStructuringElement(cv2.MORPH_RECT, (3, 3))
-            m = cv2.morphologyEx(m, cv2.MORPH_OPEN, k0, iterations=1)
-        k1 = cv2.getStructuringElement(cv2.MORPH_RECT, (_odd(ck), _odd(ck)))
-        m = cv2.morphologyEx(m, cv2.MORPH_CLOSE, k1, iterations=1)
-        k2 = cv2.getStructuringElement(cv2.MORPH_RECT, (_odd(dk), _odd(dk)))
-        m = cv2.dilate(m, k2, iterations=1)
-        return m
-
-    def _ar_score(self, ar: float) -> float:
-        lo, hi = self.geo.ar_min, self.geo.ar_max
-        mid = 1.0
-        span = max(mid - lo, hi - mid)
-        return float(max(0.0, 1.0 - abs(ar - mid) / span))
-
-    def _shape_features(self, cnt: NDArray, W: int, H: int) -> Dict[str, Any]:
-        area = cv2.contourArea(cnt)
-        per = max(1e-6, cv2.arcLength(cnt, True))
-        x, y, w, h = cv2.boundingRect(cnt)
-        hull = cv2.convexHull(cnt)
-        a_hull = max(1e-6, cv2.contourArea(hull))
-        solidity = float(area / a_hull)
-        circular = float(min(1.0, 4.0 * np.pi * area / (per * per)))
-        rectangularity = float(area / (w * h))
-        ar = w / max(1.0, h)
-        ar_s = self._ar_score(ar)
-        bbox_ratio = (w * h) / (W * H)
-        fill = rectangularity
-        return {
-            "area": area, "per": per, "bbox": (x, y, w, h), "ar": ar,
-            "solidity": solidity, "circular": circular, "rect": rectangularity,
-            "ar_s": ar_s, "bbox_ratio": bbox_ratio, "fill": fill
-        }
-
-    def _score_contour(self, feat: Dict[str, Any], cx_img: float, cy_img: float, W: int, H: int):
-        x, y, w, h = feat["bbox"]
-        area_norm = feat["area"] / (W * H)
-        cx = x + w / 2.0
-        cy = y + h / 2.0
-        dist = np.hypot(cx - cx_img, cy - cy_img) / np.hypot(cx_img, cy_img)
-        sc = (
-            self.w.area * area_norm +
-            self.w.fill * feat["fill"] +
-            self.w.solidity * feat["solidity"] +
-            self.w.circular * feat["circular"] +
-            self.w.rect * feat["rect"] +
-            self.w.ar * feat["ar_s"] -
-            (self.w.dist * self.w.center_bias) * dist
-        )
-        return float(sc), float(dist)
-
-    def _select_best(self, mask: NDArray, min_area_px: int, W: int, H: int, hard_cap: float):
-        cnts, _ = cv2.findContours(mask, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
-        if not cnts:
-            return None
-        cx_img, cy_img = W / 2.0, H / 2.0
-        best, best_s = None, -1e9
-        for c in cnts:
-            a = cv2.contourArea(c)
-            if a < min_area_px:
-                continue
-            x, y, w, h = cv2.boundingRect(c)
-            ar = w / max(1.0, h)
-            bbox_ratio = (w * h) / (W * H)
-            if not (self.geo.ar_min <= ar <= self.geo.ar_max):
-                continue
-            if bbox_ratio > hard_cap:
-                continue
-            feat = self._shape_features(c, W, H)
-            s, _ = self._score_contour(feat, cx_img, cy_img, W, H)
-            if s > best_s:
-                best_s = s
-                best = dict(cnt=c, score=s, **feat)
-        return best
-
-    def _process_with_margin(self, edges: NDArray, margin: int):
-        e = edges.copy()
-        if margin > 0:
-            e[:margin, :] = 0
-            e[-margin:, :] = 0
-            e[:, :margin] = 0
-            e[:, -margin:] = 0
-
-        H, W = e.shape[:2]
-        min_area_px = int(self.geo.min_area_frac * W * H)
-        ck, dk = 3, 3
-        opening = False
-        best = None
-
-        for _ in range(1, self.morph_cfg.steps + 1):
-            m = self._run_morph(e, ck, dk, opening=opening)
-            info = self._select_best(m, min_area_px, W, H, self.geo.bbox_hard_cap)
-            if info is None:
-                ck = min(self.morph_cfg.close_max, ck + 2)
-                dk = min(self.morph_cfg.dil_max, dk + 2)
-                opening = True
-                continue
-
-            best = (m, info, ck, dk)
-            if (self.geo.bbox_min <= info["bbox_ratio"] <= self.geo.bbox_max) and (self.geo.fill_min <= info["fill"] <= self.geo.fill_max):
-                break
-            if (info["fill"] < self.geo.fill_min) or (info["bbox_ratio"] < self.geo.bbox_min):
-                ck = min(self.morph_cfg.close_max, ck + 2)
-                dk = min(self.morph_cfg.dil_max, dk + 2)
-            elif (info["fill"] > self.geo.fill_max) or (info["bbox_ratio"] > self.geo.bbox_max):
-                dk = max(self.morph_cfg.dil_min, dk - 2)
-                ck = max(self.morph_cfg.close_min, ck - 1)
-                opening = True
-
-        return best, e
-
-    def _try_with_margins(self, edges: NDArray):
-        best, e_used = self._process_with_margin(edges, self.proc.border_margin)
-        if best is None:
-            best, e_used = self._process_with_margin(edges, 0)
-        return best, e_used
-
-    def _draw_overlay(self, proc: NDArray, info: Dict[str, Any], mask_final: NDArray):
-        overlay = proc.copy()
-        x, y, w, h = info["bbox"]
-        cv2.drawContours(mask_final, [info["cnt"]], -1, 255, thickness=cv2.FILLED)
-        cv2.rectangle(overlay, (x, y), (x + w, y + h), (0, 255, 0), 2)
-        M = cv2.moments(info["cnt"])
-        c = (x + w // 2, y + h // 2)
-        if M["m00"] != 0:
-            c = (int(M["m10"] / M["m00"]), int(M["m01"] / M["m00"]))
-        cv2.circle(overlay, c, 4, (0, 255, 0), -1)
-        tag = "color_gate" if self.color.enabled else "canny"
-        txt = f"{tag}  fill={info['fill']:.2f}  bbox={info['bbox_ratio']:.2f}  sc={info['score']:.2f}"
-        cv2.putText(overlay, txt, (x, max(18, y - 6)), cv2.FONT_HERSHEY_SIMPLEX, 0.5, (0, 255, 0), 2)
-        return overlay, c
-
-    # -------------- color gate --------------
-    def _color_gate(self, bgr: NDArray) -> NDArray:
-        if self.color.mode == "hsv":
-            hsv = cv2.cvtColor(bgr, cv2.COLOR_BGR2HSV)
-            lo = np.array(self.color.hsv_lo, dtype=np.uint8)
-            hi = np.array(self.color.hsv_hi, dtype=np.uint8)
-            mask = cv2.inRange(hsv, lo, hi)
-            return mask
-        # default: lab_bg
-        lab = cv2.cvtColor(bgr, cv2.COLOR_BGR2LAB)
-        a = lab[:,:,1].astype(np.float32)
-        b = lab[:,:,2].astype(np.float32)
-        a0 = float(np.median(a))
-        b0 = float(np.median(b))
-        dist = np.sqrt((a - a0)**2 + (b - b0)**2)
-        mask = (dist > float(self.color.ab_thresh)).astype(np.uint8) * 255
-        return mask
 
 # ----------------------- CLI helper -----------------------
 def run_file(image_path: str, profile: Optional[str] = None, out_dir: Optional[str] = "results"):
